Refactoring execution result structure.
Now, outputs are `dict[str, data]` where the keys are the output
series. Losses are also dict, where keys are the loss names.
Also, execution result has a size, according to which the losses
are weighted during reduction in `join_execution_results` (a method
which has also been refactored).

Generic trainer is therefore now a feedable, because it needs the batch
size for its execution result.